feat(event-sorting): Created table entry types & renamed event type file
