+++ apps/website/src/types/account.d.ts
@@ -1,0 +1,10 @@
+export interface AccountTableEntry {
+  name: string;
+  tag: string;
+  townhall: number;
+  clan: {
+    id: number;
+    name: string;
+    tag: string;
+  };
+}
