--- apps/website/src/types/event-status.ts
@@ -1,5 +1,0 @@
-export enum EventStatus {
-  Ongoing = "Ongoing",
-  Finished = "Finished",
-  Upcoming = "Upcoming",
-}
