+++ apps/website/src/types/event.d.ts
@@ -1,0 +1,15 @@
+export enum EventStatus {
+  Ongoing = "Ongoing",
+  Finished = "Finished",
+  Upcoming = "Upcoming",
+}
+
+export interface EventTableEntry {
+  id: number;
+  status: EventStatus;
+  name: string;
+  isGift: boolean;
+  startDate: Date;
+  endDate: Date;
+  maxChests: number;
+}
